20/5/25
I finally got around to finishing me documentations (DESIGN). I also removed all of my testing because I found out I did not need it for Sprint 1. I also completed a prototype of my game which shows the basic movement and player stats display. I also created a a bunch of UI functions which allows me to create the rest of the program a lot easier. I hope to complete the review and launch and move on to Sprint 2 by the end of this week.