17/6/25
I added onto the storyline, you can now go into a village and buy potions. The village also restores your health and mana to 100%. I also added onto the combat system, making it so that you can use potions as well as equip items. I also added a feature that prints out the storyline slowly, making it easier to read and understand. That was all that was added. I just need to do the requirements and readme as well as reviewing my current sprint then its onto sprint 4!