19/6/25
Finished practical. The practical is all complete, with the addition of a revamped inventory systems, addition to the storyline, completing it as well as small changes to the Combat System. I will now remake the Structure chart to include the combat system and its interactions with the storyline and UI. Afterwards, I will review my project and launch it.